Fixes bug in handling of too large file uploads
The commons-fileupload implementation detects too large uploads
already when reading the content size from the request header.
The resulting exception was thrown before any uploads, on which the
STATUS_ERROR_OVERSIZE could be set, were registered in the upload processor.
This fix creates a dummy upload object in this case on which the status can
be set and with which the listeners can be notified.